v1.1.0
The major update is that the payload length is encrypted too. This
should prevent attacks from the large payload length attach. Because
of the payload length is not encrypted, the attacker could send a handshake
packet with the first 4-byte was set to a large integer to the
remote-proxy to cause the remote-proxy OOM.